Reorganized project; moved logic out of ProjectPicker
