+++ Lobe_iOS/Views/OpenScreen/ProjectRow.swift
